feat: implement truck class
This commit introduces the truck class to be used by both Scania
and CarTransport. The issue with this implementation is that while
they both require a bed, the CarTransport is binary (up/down), while
the Scania has an angle. This results in the Truck class being pretty
bare-bones.

Another solution would be to handle the angles in the Truck class,
and then in CarTransport set an arbitrary angle as the up state
and another as the down, and make it binary in that way - however,
I think that is an ugly solution.